pyproject.toml
[tool.poetry]
name = "adm-simcc"
version = "1.0.0"
description = ""
authors = ["meirelescg <geu_costa@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.2,<3.9.7 || >3.9.7,<4.0.0"
flask = "^3.0.3"
flask-cors = "^4.0.0"
pandas = "^2.2.2"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.8.2"}
httpie = "^3.2.2"
zeep = "^4.2.1"
streamlit = "^1.35.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"



[tool.poetry.group.dev.dependencies]
testcontainers = {extras = ["postgres"], version = "^4.5.1"}
factory-boy = "^3.3.0"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
ruff = "^0.4.3"
httpx = "^0.27.0"
pytest-flask = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 80

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.pytest.ini_options]
addopts = '-p no:warnings'

[tool.taskipy.tasks]
run = 'gunicorn --certfile=$CERT_FILE --keyfile=$KEY_FILE -b 0.0.0.0:5000 adm_simcc:create_app --reload'
test = 'pytest -x -s --cov=adm_simcc/rest -vv --new-first'
post_test = 'coverage html'
